refactor: extract classes into separate files for better organization
- Extract ExtensionDependencies, RemoteEnvironmentHandler, and ExtensionInitializer from extension.ts
- Extract all tree item classes from workspacesProvider.ts to workspacesProvider/treeItems.ts
- Follow TypeScript camelCase naming conventions for files and directories
- Improve code organization and maintainability by following single responsibility principle
