+++ src/extension.ts
@@ -1,99 +1,14 @@
 "use strict";
-import axios, { isAxiosError } from "axios";
-import { getErrorMessage } from "coder/site/src/api/errors";
 import * as module from "module";
 import * as vscode from "vscode";
 import { makeCoderSdk, needToken } from "./api";
-import { errToStr } from "./api-helper";
 import { Commands } from "./commands";
-import { getErrorDetail } from "./error";
+import { ExtensionDependencies } from "./extension/dependencies";
+import { ExtensionInitializer } from "./extension/initializer";
 import { Logger } from "./logger";
-import { Remote } from "./remote";
 import { Storage } from "./storage";
-import { DefaultUIProvider } from "./uiProvider";
 import { toSafeHost } from "./util";
 import { WorkspaceQuery, WorkspaceProvider } from "./workspacesProvider";
-
-class ExtensionDependencies {
-	public readonly vscodeProposed: typeof vscode;
-	public readonly remoteSSHExtension: vscode.Extension<unknown> | undefined;
-	public readonly output: vscode.OutputChannel;
-	public readonly storage: Storage;
-	public readonly logger: Logger;
-	public readonly restClient: ReturnType<typeof makeCoderSdk>;
-	public readonly uiProvider: DefaultUIProvider;
-	public readonly commands: Commands;
-	public readonly myWorkspacesProvider: WorkspaceProvider;
-	public readonly allWorkspacesProvider: WorkspaceProvider;
-
-	private constructor(
-		vscodeProposed: typeof vscode,
-		remoteSSHExtension: vscode.Extension<unknown> | undefined,
-		output: vscode.OutputChannel,
-		storage: Storage,
-		logger: Logger,
-		restClient: ReturnType<typeof makeCoderSdk>,
-		uiProvider: DefaultUIProvider,
-		commands: Commands,
-		myWorkspacesProvider: WorkspaceProvider,
-		allWorkspacesProvider: WorkspaceProvider,
-	) {
-		this.vscodeProposed = vscodeProposed;
-		this.remoteSSHExtension = remoteSSHExtension;
-		this.output = output;
-		this.storage = storage;
-		this.logger = logger;
-		this.restClient = restClient;
-		this.uiProvider = uiProvider;
-		this.commands = commands;
-		this.myWorkspacesProvider = myWorkspacesProvider;
-		this.allWorkspacesProvider = allWorkspacesProvider;
-	}
-
-	static async create(
-		ctx: vscode.ExtensionContext,
-	): Promise<ExtensionDependencies> {
-		// Setup remote SSH extension
-		const { vscodeProposed, remoteSSHExtension } = setupRemoteSSHExtension();
-
-		// Create output channel
-		const output = vscode.window.createOutputChannel("Coder");
-
-		// Initialize infrastructure
-		const { storage, logger } = await initializeInfrastructure(ctx, output);
-
-		// Initialize REST client
-		const restClient = await initializeRestClient(storage);
-
-		// Setup tree views
-		const { myWorkspacesProvider, allWorkspacesProvider } = setupTreeViews(
-			restClient,
-			storage,
-		);
-
-		// Create UI provider and commands
-		const uiProvider = new DefaultUIProvider(vscodeProposed.window);
-		const commands = new Commands(
-			vscodeProposed,
-			restClient,
-			storage,
-			uiProvider,
-		);
-
-		return new ExtensionDependencies(
-			vscodeProposed,
-			remoteSSHExtension,
-			output,
-			storage,
-			logger,
-			restClient,
-			uiProvider,
-			commands,
-			myWorkspacesProvider,
-			allWorkspacesProvider,
-		);
-	}
-}
 
 export function setupRemoteSSHExtension(): {
 	vscodeProposed: typeof vscode;
@@ -365,153 +280,6 @@ export function registerCommands(
 		"coder.viewLogs",
 		commands.viewLogs.bind(commands),
 	);
-}
-
-class RemoteEnvironmentHandler {
-	private readonly vscodeProposed: typeof vscode;
-	private readonly remoteSSHExtension: vscode.Extension<unknown> | undefined;
-	private readonly restClient: ReturnType<typeof makeCoderSdk>;
-	private readonly storage: Storage;
-	private readonly commands: Commands;
-	private readonly extensionMode: vscode.ExtensionMode;
-
-	constructor(
-		deps: ExtensionDependencies,
-		extensionMode: vscode.ExtensionMode,
-	) {
-		this.vscodeProposed = deps.vscodeProposed;
-		this.remoteSSHExtension = deps.remoteSSHExtension;
-		this.restClient = deps.restClient;
-		this.storage = deps.storage;
-		this.commands = deps.commands;
-		this.extensionMode = extensionMode;
-	}
-
-	async initialize(): Promise<boolean> {
-		// Skip if no remote SSH extension or no remote authority
-		if (!this.remoteSSHExtension || !this.vscodeProposed.env.remoteAuthority) {
-			return true; // No remote environment to handle
-		}
-
-		const remote = new Remote(
-			this.vscodeProposed,
-			this.storage,
-			this.commands,
-			this.extensionMode,
-		);
-
-		try {
-			const details = await remote.setup(
-				this.vscodeProposed.env.remoteAuthority,
-			);
-			if (details) {
-				// Authenticate the plugin client
-				this.restClient.setHost(details.url);
-				this.restClient.setSessionToken(details.token);
-			}
-			return true; // Success
-		} catch (ex) {
-			await this.handleRemoteError(ex);
-			// Always close remote session when we fail to open a workspace
-			await remote.closeRemote();
-			return false; // Failed
-		}
-	}
-
-	private async handleRemoteError(error: unknown): Promise<void> {
-		if (
-			error &&
-			typeof error === "object" &&
-			"x509Err" in error &&
-			"showModal" in error
-		) {
-			const certError = error as {
-				x509Err?: string;
-				message?: string;
-				showModal: (title: string) => Promise<void>;
-			};
-			this.storage.writeToCoderOutputChannel(
-				certError.x509Err || certError.message || "Certificate error",
-			);
-			await certError.showModal("Failed to open workspace");
-		} else if (isAxiosError(error)) {
-			const msg = getErrorMessage(error, "None");
-			const detail = getErrorDetail(error) || "None";
-			const urlString = axios.getUri(error.config);
-			const method = error.config?.method?.toUpperCase() || "request";
-			const status = error.response?.status || "None";
-			const message = `API ${method} to '${urlString}' failed.\nStatus code: ${status}\nMessage: ${msg}\nDetail: ${detail}`;
-			this.storage.writeToCoderOutputChannel(message);
-			await this.vscodeProposed.window.showErrorMessage(
-				"Failed to open workspace",
-				{
-					detail: message,
-					modal: true,
-					useCustom: true,
-				},
-			);
-		} else {
-			const message = errToStr(error, "No error message was provided");
-			this.storage.writeToCoderOutputChannel(message);
-			await this.vscodeProposed.window.showErrorMessage(
-				"Failed to open workspace",
-				{
-					detail: message,
-					modal: true,
-					useCustom: true,
-				},
-			);
-		}
-	}
-}
-
-class ExtensionInitializer {
-	private readonly deps: ExtensionDependencies;
-	private readonly ctx: vscode.ExtensionContext;
-
-	constructor(deps: ExtensionDependencies, ctx: vscode.ExtensionContext) {
-		this.deps = deps;
-		this.ctx = ctx;
-	}
-
-	async initialize(): Promise<void> {
-		// Register URI handler and commands
-		this.registerHandlers();
-
-		// Handle remote environment if applicable
-		const remoteHandler = new RemoteEnvironmentHandler(
-			this.deps,
-			this.ctx.extensionMode,
-		);
-		const remoteHandled = await remoteHandler.initialize();
-		if (!remoteHandled) {
-			return; // Exit early if remote setup failed
-		}
-
-		// Initialize authentication
-		await initializeAuthentication(
-			this.deps.restClient,
-			this.deps.storage,
-			this.deps.myWorkspacesProvider,
-			this.deps.allWorkspacesProvider,
-		);
-	}
-
-	private registerHandlers(): void {
-		// Register URI handler
-		registerUriHandler(
-			this.deps.commands,
-			this.deps.restClient,
-			this.deps.storage,
-		);
-
-		// Register commands
-		registerCommands(
-			this.deps.commands,
-			this.deps.myWorkspacesProvider,
-			this.deps.allWorkspacesProvider,
-		);
-	}
 }
 
 export async function initializeAuthentication(
